scripts/requirements.txt
supabase==2.3.0
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.11
geojson==3.1.0
python-dateutil==2.9.0
